+++ packages/ragbits-evaluate/src/ragbits/evaluate/config.py
@@ -9,5 +9,12 @@ class EvaluateConfig(CoreConfig):
 
     dataloader_default_class: str = "ragbits.evaluate.dataloaders.hf:HFDataLoader"
 
+    default_input_schemas_for_pipelines: dict[str, dict] = {
+        "document_search_evaluation": {
+            "type": "ragbits.evaluate.pipelines.document_search:DocumentSearchDatapointSchema",
+            "config": {"question_col": "question", "reference_passage_col": "passage"},
+        }
+    }
+
 
 eval_config = get_config_instance(EvaluateConfig, subproject="evaluate")