+++ packages/ragbits-evaluate/src/ragbits/evaluate/pipelines/document_search.py
@@ -5,7 +5,14 @@ from typing_extensions import Self
 
 from ragbits.core.sources.hf import HuggingFaceSource
 from ragbits.document_search import DocumentSearch
-from ragbits.evaluate.pipelines.base import EvaluationPipeline, EvaluationResult
+from ragbits.evaluate.pipelines.base import EvaluationDatapointSchema, EvaluationPipeline, EvaluationResult
+
+
+class DocumentSearchDatapointSchema(EvaluationDatapointSchema):
+    """A column description for document search"""
+
+    question_col: str
+    reference_passage_col: str
 
 
 @dataclass
@@ -19,10 +26,12 @@ class DocumentSearchResult(EvaluationResult):
     predicted_passages: list[str]
 
 
-class DocumentSearchPipeline(EvaluationPipeline[DocumentSearch]):
+class DocumentSearchPipeline(EvaluationPipeline[DocumentSearch, DocumentSearchDatapointSchema]):
     """
     Document search evaluation pipeline.
     """
+
+    configuration_key = "document_search_evaluation"
 
     def __init__(self, evaluation_target: DocumentSearch, source: dict | None = None) -> None:
         """
@@ -67,20 +76,21 @@ class DocumentSearchPipeline(EvaluationPipeline[DocumentSearch]):
             )
             await self.evaluation_target.ingest(sources)
 
-    async def __call__(self, data: dict) -> DocumentSearchResult:
+    async def __call__(self, data: dict, schema: DocumentSearchDatapointSchema) -> DocumentSearchResult:
         """
         Runs the document search evaluation pipeline.
 
         Args:
             data: The evaluation data.
+            schema: column names specification
 
         Returns:
             The evaluation result.
         """
-        elements = await self.evaluation_target.search(data["question"])
+        elements = await self.evaluation_target.search(data[schema.question_col])
         predicted_passages = [element.text_representation for element in elements if element.text_representation]
         return DocumentSearchResult(
-            question=data["question"],
-            reference_passages=data["passage"],
+            question=data[schema.question_col],
+            reference_passages=data[schema.reference_passage_col],
             predicted_passages=predicted_passages,
         )